Change build.sh to build.txt
This is so build.sh does not need to be signed, and still can be included as a resource. This is an attempt to fix this error: ERROR ITMS-90135: "The executable could not be re-signed for submission to the App Store. The app may have been built or signed with non-compliant or pre-release tools."  https://github.com/bibledit/cloud/issues/395